+++ src/Core/Film.h
@@ -19,26 +19,26 @@ namespace Raycer
 		~Film();
 
 		void clear();
-		void resize(uint64_t width, uint64_t height);
-		void addSample(uint64_t x, uint64_t y, const Color& color, float filterWeight);
-		void addSample(uint64_t index, const Color& color, float filterWeight);
+		void resize(uint32_t width, uint32_t height);
+		void addSample(uint32_t x, uint32_t y, const Color& color, float filterWeight);
+		void addSample(uint32_t index, const Color& color, float filterWeight);
 
-		Color getLinearColor(uint64_t x, uint64_t y) const;
-		Color getOutputColor(uint64_t x, uint64_t y) const;
+		Color getLinearColor(uint32_t x, uint32_t y) const;
+		Color getOutputColor(uint32_t x, uint32_t y) const;
 		
 		void generateImage(Tonemapper& tonemapper);
 		const Image& getImage() const;
 
-		uint64_t getWidth() const;
-		uint64_t getHeight() const;
+		uint32_t getWidth() const;
+		uint32_t getHeight() const;
 
-		std::atomic<uint64_t> pixelSamples;
+		std::atomic<uint32_t> pixelSamples;
 
 	private:
 
-		uint64_t width = 0;
-		uint64_t height = 0;
-		uint64_t length = 0;
+		uint32_t width = 0;
+		uint32_t height = 0;
+		uint32_t length = 0;
 		
 		Color* pixels = nullptr;
