+++ src/Core/Image.h
@@ -22,18 +22,18 @@ namespace Raycer
 		Image();
 		~Image();
 
-		explicit Image(uint64_t length);
-		Image(uint64_t width, uint64_t height);
-		Image(uint64_t width, uint64_t height, float* rgbaData);
+		explicit Image(uint32_t length);
+		Image(uint32_t width, uint32_t height);
+		Image(uint32_t width, uint32_t height, float* rgbaData);
 		explicit Image(const std::string& fileName);
 
-		void load(uint64_t width, uint64_t height, float* rgbaData);
+		void load(uint32_t width, uint32_t height, float* rgbaData);
 		void load(const std::string& fileName);
 		void save(const std::string& fileName, bool writeToLog = true) const;
-		void resize(uint64_t length);
-		void resize(uint64_t width, uint64_t height);
-		void setPixel(uint64_t x, uint64_t y, const Color& color);
-		void setPixel(uint64_t index, const Color& color);
+		void resize(uint32_t length);
+		void resize(uint32_t width, uint32_t height);
+		void setPixel(uint32_t x, uint32_t y, const Color& color);
+		void setPixel(uint32_t index, const Color& color);
 		void clear();
 		void clear(const Color& color);
 		void applyGamma(float gamma);
@@ -41,12 +41,12 @@ namespace Raycer
 		void swapComponents();
 		void fillWithTestPattern();
 
-		uint64_t getWidth() const;
-		uint64_t getHeight() const;
-		uint64_t getLength() const;
+		uint32_t getWidth() const;
+		uint32_t getHeight() const;
+		uint32_t getLength() const;
 
-		Color getPixel(uint64_t x, uint64_t y) const;
-		Color getPixel(uint64_t index) const;
+		Color getPixel(uint32_t x, uint32_t y) const;
+		Color getPixel(uint32_t index) const;
 		Color getPixelNearest(float u, float v) const;
 		Color getPixelBilinear(float u, float v) const;
 		Color getPixelBicubic(float u, float v, Filter& filter) const;
@@ -56,9 +56,9 @@ namespace Raycer
 
 	private:
 
-		uint64_t width = 0;
-		uint64_t height = 0;
-		uint64_t length = 0;
+		uint32_t width = 0;
+		uint32_t height = 0;
+		uint32_t length = 0;
 
 		Color* pixels = nullptr;
 	};
